CI: create utils_cpuid module
Created a module for holding auxiliary
functions for testing cpu features.
Moved there the relevant functions.
