+++ tests/framework/utils.py
@@ -10,7 +10,6 @@ import re
 import subprocess
 import threading
 import typing
-from enum import Enum, auto
 from collections import namedtuple, defaultdict
 import time
 import psutil
@@ -214,29 +213,6 @@ class StoppableThread(threading.Thread):
     def stopped(self):
         """Check if the thread was stopped."""
         return self._should_stop
-
-
-class CpuVendor(Enum):
-    """CPU vendors enum."""
-
-    AMD = auto()
-    INTEL = auto()
-
-
-def get_cpu_vendor():
-    """Return the CPU vendor."""
-    brand_str = subprocess.check_output("lscpu", shell=True).strip().decode()
-    if 'AuthenticAMD' in brand_str:
-        return CpuVendor.AMD
-    return CpuVendor.INTEL
-
-
-def get_cpu_model_name():
-    """Return the CPU model name."""
-    _, stdout, _ = run_cmd("cat /proc/cpuinfo | grep 'model name' | uniq")
-    info = stdout.strip().split(sep=":")
-    assert len(info) == 2
-    return info[1].strip()
 
 
 def search_output_from_cmd(cmd: str,
