+++ tests/framework/utils_cpuid.py
@@ -1,0 +1,31 @@
+# Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Helper functions for testing CPU identification functionality."""
+
+import subprocess
+from enum import Enum, auto
+
+from framework.utils import run_cmd
+
+
+class CpuVendor(Enum):
+    """CPU vendors enum."""
+
+    AMD = auto()
+    INTEL = auto()
+
+
+def get_cpu_vendor():
+    """Return the CPU vendor."""
+    brand_str = subprocess.check_output("lscpu", shell=True).strip().decode()
+    if 'AuthenticAMD' in brand_str:
+        return CpuVendor.AMD
+    return CpuVendor.INTEL
+
+
+def get_cpu_model_name():
+    """Return the CPU model name."""
+    _, stdout, _ = run_cmd("cat /proc/cpuinfo | grep 'model name' | uniq")
+    info = stdout.strip().split(sep=":")
+    assert len(info) == 2
+    return info[1].strip()
